Cambiando la cadenas de conexion y la subida

+++ Controllers/CervezaController.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Linq.Expressions;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -68,6 +69,122 @@ namespace WebServiceBlazorCrud.Controllers
             return Ok(cerveza);
         }
 
+        public static List<Cerveza> BusquedaAvanzada(int param, string criterio,Respuesta resp)
+        {
+            BlazorCrudContext db = new BlazorCrudContext();
+            
+            List<Cerveza> oCerveza = new List<Cerveza>();
+
+            try
+            {
+                switch (param)
+                {
+                    case 0:
+                        oCerveza = db.Cervezas.ToList();
+                        resp.Exito = 1;
+                        break;
+                    case 1:
+                        oCerveza = db.Cervezas.Where(a => a.Id == Convert.ToInt64(criterio)).ToList();
+                        resp.Exito = 1;
+                        break;
+                    case 2:
+                        oCerveza = db.Cervezas.Where(a => a.Nombre.Contains(criterio)).ToList();
+                        resp.Exito = 1;
+                        break;
+                    case 3:
+                        oCerveza = db.Cervezas.Where(a => a.Marca.Contains(criterio)).ToList();
+                        resp.Exito = 1;
+                        break;
+
+                    default:
+                        oCerveza = new List<Cerveza>();
+                        resp.Exito = 0;
+                        resp.Mensaje = "Este parametro no existe";
+                        break;
+                }
+
+
+
+            }
+            catch (Exception ex)
+            {
+                resp.Mensaje = ex.Message;
+              
+            }
+            finally
+            {
+                db.Dispose();
+            }
+
+
+            return oCerveza;
+        }
+
+        public static KeyValuePair<Respuesta, List<Cerveza>> GetValores(int param, string criterio)
+        {
+            BlazorCrudContext db = new BlazorCrudContext();
+            Respuesta resp = new Respuesta();
+            
+
+            List<Cerveza> oCerveza = new List<Cerveza>();
+
+            try
+            {
+                switch (param)
+                {
+                    case 0:
+                        oCerveza = db.Cervezas.ToList();
+                        resp.Exito = 1;
+                        break;
+                    case 1:
+                        oCerveza = db.Cervezas.Where(a => a.Id == Convert.ToInt64(criterio)).ToList();
+                        resp.Exito = 1;
+                        break;
+                    case 2:
+                        oCerveza = db.Cervezas.Where(a => a.Nombre.Contains(criterio)).ToList();
+                        resp.Exito = 1;
+                        break;
+                    case 3:
+                        oCerveza = db.Cervezas.Where(a => a.Marca.Contains(criterio)).ToList();
+                        resp.Exito = 1;
+                        break;
+
+                    default:
+                        oCerveza = new List<Cerveza>();
+                        resp.Exito = 0;
+                        resp.Mensaje = "Este parametro no existe";
+                        break;
+                }
+
+
+
+            }
+            catch (Exception ex)
+            {
+                resp.Mensaje = ex.Message;
+
+            }
+            finally
+            {
+                db.Dispose();
+            }
+
+
+            return new KeyValuePair<Respuesta, List<Cerveza>>(resp, oCerveza);
+        }
+
+        [HttpGet("{param}/{criterio}")]
+        public IActionResult GetAvanzado(int param, string criterio)
+        {
+            var valores = GetValores(param, criterio);
+            List<Cerveza> oCerveza = valores.Value;
+            Respuesta respuesta = valores.Key;
+
+            
+
+            return Ok(oCerveza);
+        }
+
         [HttpPost]
         public IActionResult Add(CervezaRequest model)
         {
@@ -106,11 +223,20 @@ namespace WebServiceBlazorCrud.Controllers
                 using (BlazorCrudContext db = new BlazorCrudContext())
                 {
                     Cerveza oCerveza = db.Cervezas.Find(model.Id);
-                    oCerveza.Nombre = model.Nombre;
-                    oCerveza.Marca = model.Marca;
-                    db.Entry(oCerveza).State = Microsoft.EntityFrameworkCore.EntityState.Modified;
-                    db.SaveChanges();
-                    oRespuesta.Exito = 1;
+                    if(oCerveza!=null)
+                    {
+                        oCerveza.Nombre = model.Nombre;
+                        oCerveza.Marca = model.Marca;
+                        db.Entry(oCerveza).State = Microsoft.EntityFrameworkCore.EntityState.Modified;
+                        db.SaveChanges();
+                        oRespuesta.Exito = 1;
+                    }
+                    else
+                    {
+                        oRespuesta.Mensaje = "La cerveza que desea eliminar no existe";
+                        oRespuesta.Exito = 0;
+                    }
+                    
                 }
             }
             catch (Exception ex)
@@ -133,9 +259,18 @@ namespace WebServiceBlazorCrud.Controllers
                 using (BlazorCrudContext db = new BlazorCrudContext())
                 {
                     Cerveza oCerveza = db.Cervezas.Find(Id);
-                    db.Remove(oCerveza);
-                    db.SaveChanges();
-                    oRespuesta.Exito = 1;
+                    if(oCerveza!=null)
+                    {
+                        db.Remove(oCerveza);
+                        db.SaveChanges();
+                        oRespuesta.Exito = 1;
+                    }
+                    else
+                    {
+                        oRespuesta.Mensaje = "El elemento que busca no pudo ser encontrado";
+                        oRespuesta.Exito = 0;
+                    }
+                    
                 }
             }
